+++ js/data-for-create-photo-descriptions.js
@@ -1,0 +1,31 @@
+const getDataCreatePhotoDescriptions = () => {
+  const COMMENTS = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце-концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как-будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+  const DESCRIPTIONS = [
+    'Мороз и солнце; день чудесный!',
+    'Навстречу северной Авроры',
+    'Вечор, ты помнишь, вьюга злилась',
+    'На мутном небе мгла носилась',
+    'Луна, как бледное пятно',
+    'Под голубыми небесами',
+    'Великолепными коврами',
+    'Блестя на солнце, снег лежит',
+    'Прозрачный лес один чернеет',
+    'И ель сквозь иней зеленеет',
+    'И речка подо льдом блестит',
+    'Вся комната янтарным блеском',
+  ];
+
+  const NAMES = ['Николай', 'Аким', 'Ким', 'Харитон', 'Тимур', 'Степан'];
+
+  return {COMMENTS, DESCRIPTIONS, NAMES};
+};
+
+export {getDataCreatePhotoDescriptions};
