--- js/zamykaniya.js
@@ -1,15 +1,0 @@
-// function createIdGenerator () {
-//   let lastGeneratedId = 0;
-
-//   return function () {
-//     lastGeneratedId += 1;
-//     return lastGeneratedId;
-//   };
-// }
-
-// const generatePhotoId = createIdGenerator();
-
-// console.log(generatePhotoId()); // 1
-// console.log(generatePhotoId()); // 2
-// console.log(generatePhotoId()); // 3
-
